+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Fix the slowness when retrieving a device via private vehicle API
 - Fix with_latest_events query that is taking too much time, used in Agency API
+- Fix saving in base with the wrong name for the battery field in device Telemetry
 
 
 0.5.9 (2019-03-28)